refactor: yml 에서 토픽명 관리하도록 변경
