+++ member/src/main/java/com/smore/member/infrastructure/kafka/config/MemberTopicProperties.java
@@ -1,4 +1,4 @@
-package com.smore.seller.infrastructure.kafka;
+package com.smore.member.infrastructure.kafka.config;
 
 import java.util.Map;
 import lombok.Getter;
@@ -6,11 +6,11 @@ import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
-@ConfigurationProperties(prefix = "seller.topic")
+@ConfigurationProperties(prefix = "member.topic")
 @Component
 @Getter
 @Setter
-public class SellerTopicProperties {
+public class MemberTopicProperties {
 
     private Map<String, String> sellerRegister;
