+++ member/src/main/resources/application-local.yml
@@ -77,8 +77,14 @@ eureka:
   client:
     enabled: false
 
-topic:
-  produce:
+seller:
+  topic:
+    seller-register:
+      v1: seller.register.v1
+      v2: seller.register.v2
+
+member:
+  topic:
     seller-register:
       v1: seller.register.v1
       v2: seller.register.v2
